[VOID] ImageProcessor: Use TBB

+++ src/VoidCore/Media/Filesystem.cpp
@@ -280,7 +280,7 @@ std::string MEntry::ResolvedPath(v_frame_t frame) const
         // something like /path/to/output/file.####.ext but the range specified is 9000 - 11000
         // so the padding till frame 9999 would be 4 but on frame 10000 and above the padding would be 5, which is still valid
         // in terms of writing, in terms of reading, it might not be detected correctly everywhere
-        int padding = std::max(frame == 0 ? 1 : static_cast<int>(std::log10(frame)) + 1, m_FramePadding);
+        int padding = std::max(frame == 0 ? 1 : static_cast<int>(std::log10(frame)) + 1, static_cast<int>(m_FramePadding));
         path.reserve(m_Basepath.size() + 1 + m_Name.size() + 1 + padding + 1 + m_Extension.size());
 
         path += m_Basepath;
